# swagger.yml
openapi: 3.0.3
info:
  title: Systeme de Pointage API
  description: Documentation des endpoints disponibles pour l'authentification, le pointage, le pilotage des cours et la gestion academique.
  version: 1.1.0
servers:
  - url: http://127.0.0.1:5000
    description: Serveur local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    kioskKeyAuth:
      type: apiKey
      in: header
      name: X-Attendance-Kiosk-Key
  schemas:
...
    CheckinRequest:
      type: object
      required: [course_id, selfie]
      properties:
        course_id:
          type: integer
        selfie:
          type: string
          format: binary
        gps_lat:
          type: number
        gps_lng:
          type: number
        timestamp:
          type: string
          format: date-time
...
paths:
...
  /api/v1/etudiant/courses:
    get:
      tags: [Etudiant]
      summary: Lister les cours de la promotion de l'etudiant
      security:
        - bearerAuth: []
      parameters:
        - name: statut
          in: query
          schema:
            type: string
        - name: periode
          in: query
          schema:
            type: string
            enum: [upcoming, past]
        - name: page
          in: query
          schema:
            type: integer
        - name: per_page
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Liste des cours
  /api/v1/etudiant/attendance/checkin:
    post:
      tags: [Etudiant]
      summary: Pointer sa presence par reconnaissance faciale
      deprecated: true
      description: >-
        Deprecie (decision securite) : un etudiant connecte sur son propre appareil
        pourrait importer une ancienne photo au lieu de se scanner en direct. Le
        pointage nominal se fait desormais exclusivement via la borne securisee
        (POST /attendance/kiosk/scan puis POST /attendance/kiosk/checkin). Cette route
        reste fonctionnelle pour d'eventuelles integrations futures et renvoie
        l'en-tete Deprecation: true.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CheckinRequest'
      responses:
        '200':
          description: Pointage invalide enregistre
        '201':
          description: Pointage valide enregistre
        '400':
          description: Requete invalide
        '409':
          description: Pointage deja enregistre
        '503':
          description: Service de reconnaissance faciale indisponible ou non configure
  /api/v1/etudiant/attendance/checkin-qr:
    post:
      tags: [Etudiant]
      summary: Pointer sa presence en scannant le QR affiche en salle
      description: >-
        Methode alternative a la reconnaissance faciale : l'etudiant scanne, depuis son
        propre telephone connecte a son compte, le QR de cours affiche par la borne/le
        professeur (GET /professeur/courses/{cours_id}/qr). Le token embarque le cours_id
        et expire 120 secondes apres generation. Aucun selfie requis ici : la preuve
        d'identite vient du JWT etudiant, la preuve de presence vient a la fois de la
        fraicheur du QR et du reseau. Comme les endpoints /attendance/kiosk/*, cette
        route est restreinte au reseau Wi-Fi dedie de l'etablissement
        (KIOSK_ALLOWED_NETWORKS) : l'etudiant doit y etre connecte, une connexion en
        donnees mobiles est refusee meme avec un JWT valide et un QR non expire.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                gps_lat:
                  type: number
                gps_lng:
                  type: number
      responses:
        '200':
          description: Pointage enregistre (retard)
        '201':
          description: Pointage enregistre (present)
        '400':
          description: Requete invalide, QR invalide/expire, ou hors delai
        '403':
          description: >-
            Etudiant hors de la promotion du cours, ou requete hors du reseau Wi-Fi
            dedie (kiosk_network_forbidden)
        '404':
          description: Cours ou etudiant introuvable
        '409':
          description: Pointage deja enregistre
  /api/v1/etudiant/attendance/history:
    get:
      tags: [Etudiant]
      summary: Consulter l'historique des pointages
      security:
        - bearerAuth: []
      parameters:
        - name: statut
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: per_page
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Historique retourne
  /api/v1/etudiant/courses/{cours_id}/attendance/history:
    get:
      tags: [Etudiant]
      summary: Consulter l historique des pointages de l etudiant pour un cours donne
      security:
        - bearerAuth: []
      parameters:
        - name: cours_id
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: per_page
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Historique du cours retourne
        '403':
          description: Cours hors promotion de l etudiant
  /api/v1/etudiant/absences/summary:
    get:
      tags: [Etudiant]
      summary: Consulter le suivi des absences
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Synthese du suivi des absences
  /api/v1/etudiant/notifications:
    get:
      tags: [Etudiant]
      summary: Lister les notifications de l'etudiant
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: per_page
          in: query
          schema:
            type: integer
      responses:
        '200':
          description: Notifications retournees
  /api/v1/etudiant/notifications/{notification_id}/read:
    patch:
      tags: [Etudiant]
      summary: Marquer une notification comme lue
      security:
        - bearerAuth: []
      parameters:
        - name: notification_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Notification mise a jour
        '404':
          description: Notification introuvable
...